Added DTOs for degrees and degree validation results to make it easier to transfer data to and from the front end.
Added a controller which is an api that allows for course validation.

Added a DegreeService which converts Degree DTOS into their actual backend degree models.

Added an empty course validation service, which will include the logic for prerequisite checking.

Reduced some boiler plate getter and setter code with Lombok in the Course model.